+++ packages/apollo-angular-components/projects/angular/a11y/src/public_api.ts
@@ -1,1 +1,2 @@
 export { QueuedAnnouncer } from './queued-announcer/queued-announcer';
+export { UiAutoAccessibleLabelModule } from './ui-auto-accessible-label/ui-auto-accessible-label.module';