+++ packages/apollo-angular-components/projects/angular/components/ui-grid/src/components/ui-grid-toggle-columns/ui-grid-toggle-columns.module.ts
@@ -4,6 +4,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSelectModule } from '@angular/material/select';
 import { MatTooltipModule } from '@angular/material/tooltip';
+import { UiAutoAccessibleLabelModule } from '@uipath/angular/a11y';
 
 import { UiGridToggleColumnsComponent } from './ui-grid-toggle-columns.component';
 
@@ -14,6 +15,7 @@ import { UiGridToggleColumnsComponent } from './ui-grid-toggle-columns.component
         MatSelectModule,
         MatButtonModule,
         MatTooltipModule,
+        UiAutoAccessibleLabelModule,
     ],
     declarations: [UiGridToggleColumnsComponent],
     exports: [UiGridToggleColumnsComponent],